subclass `AppModDownloader`
move all GUI stuff there and keep `ModDownloader` maximally generic
