+++ src/factorio_mod_downloader/mod_downloader/__init__.py
@@ -1,0 +1,3 @@
+from .mod_downloader import ModDownloader, WebsiteDownException
+
+__all__ = ["ModDownloader", "WebsiteDownException"]
